Updated folder/file structed + moved all files to root keeing availability files in its own folder
